requirements.txt
django
djangorestframework 
django-cors-headers 
python-magic-bin
pillow
django-cors-headers
pypdf 
python-dotenv
openai
google-generativeai
beautifulsoup4 
lxml
